KTO-1060: Refaktoroitu SoraKuvausSelect ja useSoraKuvaukset
- Siirretty useSoraKuvaukset samaan tiedostoon getSoraKuvaukset-funktion kanssa ja muutett käyttämään react-querya.
- Poistettu turha useSoraKuvausOptions ja käytetty suoraan useEntityOptions-hookkia SoraKuvausSelect-komponentissa
