--- src/main/app/src/hooks/useSoraKuvaukset/index.ts
@@ -1,14 +1,0 @@
-import useApiAsync from '#/src/hooks/useApiAsync';
-import getSoraKuvaukset from '#/src/utils/soraKuvaus/getSoraKuvaukset';
-
-export const useSoraKuvaukset = ({ organisaatioOid }) => {
-  const { data: soraKuvaukset, ...rest } = useApiAsync({
-    promiseFn: getSoraKuvaukset,
-    organisaatioOid,
-    watch: organisaatioOid,
-  });
-
-  return { soraKuvaukset, ...rest };
-};
-
-export default useSoraKuvaukset;
