+++ src/main/app/src/utils/soraKuvaus/getSoraKuvaukset.ts
@@ -1,3 +1,5 @@
+import { useApiQuery } from '#/src/hooks/useApiQuery';
+
 const getSoraKuvaukset = async ({ httpClient, apiUrls, organisaatioOid }) => {
   const { data } = await httpClient.get(
     apiUrls.url('kouta-backend.soraKuvaus-list'),
@@ -5,6 +7,20 @@ const getSoraKuvaukset = async ({ httpClient, apiUrls, organisaatioOid }) => {
   );
 
   return data;
+};
+
+export const useSoraKuvaukset = ({ organisaatioOid }) => {
+  const { data: soraKuvaukset, ...rest } = useApiQuery(
+    'getSoraKuvaukset',
+    getSoraKuvaukset,
+    {
+      promiseFn: getSoraKuvaukset,
+      organisaatioOid,
+    },
+    { enabled: Boolean(organisaatioOid) }
+  );
+
+  return { soraKuvaukset, ...rest };
 };
 
 export default getSoraKuvaukset;
